requirements.txt
pandas 
requests
python-dotenv
beautifulsoup4